safari-534.53-branch/LayoutTests/media/video-src-none-expected.txt

 

Network state should remain in NETWORK_EMPTY with empty or missing 'src' attribute.
** <video> with no src attribute**
EXPECTED (videos[0].error == 'null') OK
EXPECTED (videos[0].networkState == '0') OK
EXPECTED (videos[0].src == '') OK
** <video> with empty src attribute **
EXPECTED (videos[1].error == 'null') OK
EXPECTED (videos[1].networkState == '0') OK
EXPECTED (relativeURL(videos[1].src) == 'video-src-none.html') OK

Calling load() with empty 'src' should fire 'error' event, set network state to NETWORK_NO_SOURCE, and set error to MEDIA_ERR_SRC_NOT_SUPPORTED.
'error' event:
EXPECTED (state == 'load() with empty 'src'') OK
EXPECTED ([object HTMLVideoElement] == '[object HTMLVideoElement]') OK
EXPECTED (videos[1].error.code == '4') OK
EXPECTED (videos[1].networkState == '3') OK

Calling load() with no 'src' should NOT fire 'error' event, set network state to NETWORK_EMPTY.
EXPECTED (state == 'load() with missing 'src'') OK
EXPECTED (videos[0].error == 'null') OK
EXPECTED (videos[0].networkState == '0') OK
EXPECTED (videos[0].src == '') OK
END OF TEST
